Create LoginController to handle /login endpoint for authentication
